XW-1300 | Fix recommendation thumbnails, move common logic to mixin
