+++ front/main/app/styles/component/index.scss
@@ -39,6 +39,7 @@
 @import 'article-media-linked-gallery';
 @import 'article-media-thumbnail';
 @import 'article-media-map-thumbnail';
+@import 'content-recommendation-thumbnail';
 @import 'infobox-builder';
 @import 'recent-change';
 @import 'article-diff';
